add authorization in notesApp

+++ Belajar Fundamental Back-End dengan JavaScript/Hapi Plugin dan Data Validation/Note-App/src/api/notes/handler.js
@@ -15,8 +15,14 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({
+      title,
+      body,
+      tags,
+      owner: credentialId
+    });
 
     const response = h.response({
       status: 'success',
@@ -30,7 +36,8 @@ class NotesHandler {
   }
 
   async getNotesHandler() {
-    const notes = await this._service.getNotes();
+    const { id, credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -41,6 +48,10 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(credentialId);
+
     const note = await this._service.getNoteById(id);
 
     return {
@@ -55,6 +66,9 @@ class NotesHandler {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
 
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(credentialId);
     await this._service.editNoteById(id, request.payload);
 
     return {
@@ -65,6 +79,9 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(credentialId);
     await this._service.deleteNoteById(id);
 
     return {
